recommender_system/requirements.txt
python-dotenv
pandas
groq
pinecone
fastapi 
uvicorn